+++ src/main/java/io/github/ImpactDevelopment/installer/impact/ImpactJsonVersion.java
@@ -18,16 +18,13 @@
 package io.github.ImpactDevelopment.installer.impact;
 
 import io.github.ImpactDevelopment.installer.libraries.ILibrary;
-import io.github.ImpactDevelopment.installer.libraries.LibraryBaritoneReleased;
 import io.github.ImpactDevelopment.installer.libraries.LibraryBaritoneSpecific;
 import io.github.ImpactDevelopment.installer.libraries.LibraryMaven;
 import io.github.ImpactDevelopment.installer.setting.InstallationConfig;
-import io.github.ImpactDevelopment.installer.setting.settings.BaritoneVersionSetting;
 import io.github.ImpactDevelopment.installer.setting.settings.ImpactVersionSetting;
 
 import java.util.Arrays;
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -60,14 +57,9 @@ public class ImpactJsonVersion {
                 throw new IllegalStateException("malformed " + lib.name);
             }
             String name = parts[1];
-            if (name.equals(LibraryBaritoneReleased.VARIANT)) {
-                LibraryBaritoneReleased configured = config.getSettingValue(BaritoneVersionSetting.INSTANCE);
-                boolean filtered = parts[2].contains("*");
-                if (filtered ^ configured != null) {
-                    throw new IllegalStateException(parts[2] + " " + configured + " " + baritoneVersionFilter());
-                }
-                if (configured != null) {
-                    return configured;
+            if (name.equals(LibraryBaritoneSpecific.VARIANT)) {
+                if (parts[2].contains("*")) {
+                    throw new IllegalStateException(parts[2]);
                 }
                 return new LibraryBaritoneSpecific(lib);
             }
@@ -76,16 +68,5 @@ public class ImpactJsonVersion {
             }
             return new LibraryMaven(lib);
         }).collect(Collectors.toList());
-    }
-
-    public Optional<String> baritoneVersionFilter() {
-        return Stream
-                .of(libraries)
-                .map(lib -> lib.name)
-                .map(name -> name.split(":"))
-                .filter(parts -> parts[1].equals(LibraryBaritoneReleased.VARIANT))
-                .map(parts -> parts[2])
-                .filter(filter -> filter.contains("*")) // if there is a *, that means it's a valid filter, if not it's a specific version
-                .findFirst();
     }
 }